requirements.txt
# Core Dependencies
google-cloud-bigquery>=3.11.0
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0

# Vertex AI and Generative AI
google-cloud-aiplatform[vertexai]>=1.38.0
google-generativeai>=0.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# API Framework (for future use)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Development Dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0